# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contact book API",
    "license": {
      "name": "Apache 2.0",
      "url": "<http://www.apache.org/licenses/LICENSE-2.0.html>"
    },
    "description": "APi for contact book\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "tags": [
    {
      "name": "x",
      "description": "API for contacts"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "description": "You can get access only to children added by you",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1,
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "required": false,
            "schema": {
              "type": "number",
              "default": 10,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful contact retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "contacts"
                      ],
                      "properties": {
                        "contacts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contacts"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "description": "Creates a new contact with the provided information and photo",
        "operationId": "createNewContact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "type": "object",
            "required": [
              "userId",
              "name",
              "phoneNumber"
            ],
            "properties": {
              "userId": {
                "type": "string",
                "example": "65ca67e7ae7f10c88b598384"
              },
              "name": {
                "type": "string",
                "example": "Borys"
              },
              "phoneNumber": {
                "type": "string",
                "example": 778765431
              },
              "email": {
                "type": "string",
                "example": "exampleemail@domain.com"
              },
              "isFavourite": {
                "type": "boolean",
                "example": false
              },
              "contactType": {
                "type": "string",
                "example": "personal"
              },
              "photo": {
                "oneOf": [
                  {
                    "type": "string",
                    "example": "some-url"
                  },
                  {
                    "type": "null",
                    "example": null
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contacts"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input data)"
          },
          "401": {
            "description": "Unauthorized (missing or invalid token)"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contacts by id",
        "operationId": "getContactsById",
        "description": "You can get access only to children added by you",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful contact retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 65ca67e7ae7f10c88b598384!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update a contact by ID",
        "operationId": "getPatchContactsById",
        "description": "Update specific fields of a contact by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the contact to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "description": "Name of the contact"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "description": "Phone number of the contact"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "minLength": 3,
                    "maxLength": 20,
                    "description": "Email of the contact"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Whether the contact is a favorite"
                  },
                  "contactType": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "description": "Type of the contact"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Photo of the contact"
                  }
                },
                "example": {
                  "name": "John Doe",
                  "phoneNumber": 1234567890,
                  "email": "john.doe@example.com",
                  "isFavourite": true,
                  "contactType": "friend"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully patched contact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/contacts"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete a contact by ID",
        "operationId": "deleteContactById",
        "description": "Delete a specific contact by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the contact to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted contact"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contacts": {
        "type": "object",
        "required": [
          "userId",
          "name",
          "phoneNumber"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "example": "65ca67e7ae7f10c88b598384"
          },
          "name": {
            "type": "string",
            "example": "Borys"
          },
          "phoneNumber": {
            "type": "string",
            "example": 778765431
          },
          "email": {
            "type": "string",
            "example": "exampleemail@domain.com"
          },
          "isFavourite": {
            "type": "boolean",
            "example": false
          },
          "contactType": {
            "type": "string",
            "example": "personal"
          },
          "photo": {
            "oneOf": [
              {
                "type": "string",
                "example": "some-url"
              },
              {
                "type": "null",
                "example": null
              }
            ]
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Auth header is not provided"
                },
                "data": {
                  "type": "object",
                  "example": {
                    "message": "Auth header is not provided",
                    "errorsValidation": []
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Auth header is not provided"
                },
                "data": {
                  "type": "object",
                  "example": {
                    "message": "Auth header is not provided",
                    "errorsValidation": []
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Contact not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Auth header is not provided"
                },
                "data": {
                  "type": "object",
                  "example": {
                    "message": "Auth header is not provided",
                    "errorsValidation": []
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}